+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next release
 
 - Fix a bug: fix ReservationReminderWorker, was sending reservation reminder to users with a event reservation not validated by admin + adds tests for all scenarios
+- Fix a bug: admin could not create new SupportingDocumentType (problem was on js side)
 
 ## v6.1.2 2023 October 2